refactor: project list in admin page service 이동 (admin -> project)
