+++ src/main/java/com/example/portfolio/controller/ProjectController.java
@@ -92,7 +92,7 @@ public class ProjectController {
 			@RequestParam(value = "keyWord", defaultValue = "") String keyWord) {
 		Sort sortOrder = direction.equalsIgnoreCase("asc") ? Sort.by(sort).ascending() : Sort.by(sort).descending();
 		Pageable pageable = PageRequest.of(page, size, sortOrder);
-		return adminService.getAdminProjectList(pageable, keyWord);
+		return projectService.getAdminProjectList(pageable, keyWord);
 
 	}
